Bugfix: Do not advance time_at_previous_sample_ns during GC
This was a subtle bug, which I only found due to the check for negative
`elapsed_time_ns`, since it caused time to advance to the future.

The issue was the following: if we allow
`time_at_previous_sample_ns` to advance during GC, then we can run
into cases where we double account for the same time because the period
spent in GC is added to `time_at_previous_sample` both inside
`update_time_since_previous_sample` **as well as** inside
`sample_after_gc`.

This is incorrect: the time accounting for GC should always be ignored
inside `update_time_since_previous_sample`, and always handled
separately.

Here's an example of where this could go wrong:

```
* T=00 sample      (time_at_previous_sample=0)
* T=09 on_gc_start
* T=20 on_gc_finish
* T=30 sample      (time_at_previous_sample=30)
                    <-- This is incorrect, as it should not include
                        the time spent in GC
* T=31 sample_after_gc (time_at_previous_sample=42)
                        <-- By adding 11 (time spent in gc), we've
                            now launched time_at_previous_sample
                            into the future
```

...and here's how it behaves after the fix:

```
* T=00 sample     (time_at_previous_sample=0)
* T=09 on_gc_start
* T=20 on_gc_finish
* T=30 sample     (time_at_previous_sample=9)
                   <-- This sample represents the time range [0, 9]
                       e.g. the time before GC **only**
* T=31 sample_after_gc (time_at_previous_sample=20)
                        <-- We've now (correctly) accounted for
                            the time up until on_gc_finish
```