Document intended behavior of collapsing multiple GC cycles into one
This actually came in as a bit of a bug, as it was not my original
intention, but I decided that it was actually better than my original
intention.

(The bug was that originally I intended `on_gc_finish` to not
clobber existing finish timestamps, if they were valid, but forgot
to actually do so).

Thus, the new behavior is that if the following happens:
`on_gc_start` (time=0) -> `on_gc_finish` (time=1) ->
`on_gc_start` (time=2) -> `on_gc_finish` (time=3) -> `sample_after_gc`
then we just use time=0 from the first `on_gc_start` and time=3 from the
last `on_gc_finish`, e.g. we behave as if there was a single, longer GC
period.

You'll notice that as I explained above, this did not need changing
the production code -- it already did this. What I did was clean
up the now-not-making-sense counter, as well as add specs and
comments to document what is intended.