Fix incorrect start time for profiles
**What does this PR do?**:

(Important note: This fix only affects the new CPU Profiling 2.0
profiler which is still in **alpha**, see #2209)

In #2181 we added a concurrency-safe mechanism for `StackRecorder`.
Key to how it works, is that we keep two `ddog_Profile` instances
at every point in time, and we alternate between using them.

But this alternation introduced a bug with the `start_time` of
profiles. Previously, the `start_time` got set whenever a profile got
created or reset, and we used the current time for it.

That caused the following effect:

```
t=0
  created slot_one_profile (start_time=0)
  created slow_two_profile (start_time=0)
  make_active slot_one_profile

t=60
  serialized slot_one_profile (start_time=0, finish_time=60)
  reset slot_one_profile (start_time=60)
  make_active slot_two_profile

t=120
  serialized slot_two_profile (start_time=0, finish_time=120)
  reset slot_two_profile (start_time=120)
  make_active slot_one_profile

t=180
  serialized slot_one_profile (start_time=60, finish_time=180)
  reset slow_one_profile (start_time=180)
  make_active slot_two_profile
```

That is, other than the first profile (which is why we previously
missed this bug), every profile got double the duration as intended,
because we reset it after serialization, but that profile would
not be used for the next period.

To fix this issue, we additionally change the "make_active" step
above (actually implemented in
`serializer_flip_active_and_inactive_slots`) to set the correct
`start_time` on the profile that becomes active.

Thus, we get the correct behavior:

```
t=0
  created slot_one_profile (start_time=0)
  created slow_two_profile (start_time=0) # Ignored, will be changed later
  make_active slot_one_profile (start_time=0)

t=60
  serialized slot_one_profile (start_time=0, finish_time=60)
  reset slot_one_profile (start_time=60) # Ignored, will be changed later
  make_active slot_two_profile (start_time=60) # Correct start_time

t=120
  serialized slot_two_profile (start_time=60, finish_time=120)
  reset slot_two_profile (start_time=120) # Ignored, will be changed later
  make_active slot_one_profile (start_time=120) # Correct start_time

t=180
  serialized slot_one_profile (start_time=120, finish_time=180)
  reset slow_one_profile (start_time=180)
  make_active slot_two_profile (start_time=180)
```

**Motivation**:

Having profiles with the wrong duration breaks profile aggregation.

**Additional notes**:

As I mentioned above, this only affects the new CPU Profiling 2.0
profiler codepath, so I don't expect any customers to have ever ran
into this issue.

**How to test the change?**:

Change includes coverage. Furthermore, running the profiler with
`DD_TRACE_DEBUG` prints the start/finish timestamps after
serialization, which can be used to confirm they are correct.

Finally, comparing profiles before/after in the profiler UX will
also show the difference.