Minor: Disable gc tracepoint on failure
This is not required, but it's nice -- I was doing some testing that
threw a few errors and observed that while we handled everything
correctly, the VM was still trying to invoke our tracepoint.

Since when something is wrong we want to stop the profiler as soon
as possible, it's better to instead stop the gc tracepoint immediately
as well.