[PROF-4905] Implement support for code hotspots in CpuAndWallTime collector
**What does this PR do?**:

This PR reimplements support for the code hotspots feature (connecting
traces to profiles) in the `CpuAndWallTime` collector.

Although the implementation is different, the approach is conceptually
the same as provided by `Datadog::Profiling::TraceIdentifiers::Ddtrace`
for the old profiler codepaths.

Key contraints in this implementation are:
* No objects are created to get this information
* No Ruby code is called to get this information

These constraints are key to:
* Maintain performance on the `CpuAndWallTime` collector (we never get
  out of C code).
* Keep precise control of the Global VM Lock. Since we never get out of
  C, we take the sample in one pass. Calling into Ruby code may mean
  being "interrupted" in the middle of sampling, which would break
  the internal consistency of the profiler. (Assuming that we never lose
  the GVL makes the profiler simpler and more performant).
* Correctness of customer code. Sampling happens from customer threads
  directly (using `rb_postponed_job_register_one`) so we don't want to
  call Ruby methods and potentially trigger side effects during a sample.
* Memory profiling. We don't yet implement allocation or heap
  profiling, but it's coming soon and we don't want to allocate memory
  to track memory allocations.

Thus, I chose to implement this feature by accessing the tracer context
directly from the sampling process. Because all the information needed
is available as instance variables of the objects, they can be read
safely and within the needed constraints.

This implementation also includes extensive checks to make sure that
if tracing evolves or is not available, profiling safely skips
getting the information.

Furthermore, it includes an integration test with a real tracer,
so that we'll immediately know when an incompatibility is introduced,
so that the profiler can be updated.

**Motivation**:

Code hotspots is a key feature of our APM, allowing customers to see
the profile for an individual trace.

**Additional Notes**:

This first implementation does not include support for endpoint
aggregation; this will come in a separate PR.

**How to test the change?**:

This change includes code coverage. It can also be tested by
enabling profiling and tracing on a Ruby app, and checking that
the "Code Hotspots" tab on a generated trace is showing the
profile for that trace.