Include native thread id as part of thread id label
**What does this PR do?**:

This PR extends the `thread id` label that is attached to profiling
samples to include the thread's "native thread id".

Up until now, the `thread id` that we attached was the result of
calling `Thread#object_id`. That approach allows us to easily match
what we see inside the Ruby process (because we can log the
`Thread#object_id` where something happens) with the profile.

BUT if we wanted to match the profile with what the operating system's
tooling showed, it was quite hard, because threads shown in the
OS task manager tools show an OS-level thread identifier, which is
not the same as the `Thread#object_id`.

To enable this matching, this PR changes the format of the
`thread id` from `"#{thread.object_id}"` to
`"#{thread.native_object_id} (#{thread.object_id})"` thus providing
both identifers. (This is OK with the profiling backend).

Because `Thread#native_object_id` is a Ruby 3.1+ feature, on older
Rubies we use a fallback identifier instead (`pthread_t`). This
identifier isn't as useful as the native identifier; in the future
we may want to explore a better fallback.

**Motivation**:

I found this helpful during development and when correlating
data from profiler with other tools.

**Additional Notes**:

This includes full support for macOS, even though we don't
officially support macOS.

This feature is only available on the new CPU Profiling 2.0
profiler which is still in alpha, see #2209.

**How to test the change?**:

This is easiest to test by profiling a Ruby app, and then checking
that the native ids for threads match up with what shows up on
`/proc`. (Remember, only for >= 3.1, for older Rubies it won't match.)