Workaround rspec incompatibility on Ruby 2.7 by improving test
On Ruby 2.7 and some RSpec versions, we were getting failures due to:

```
       Datadog::Tracing received :trace with unexpected arguments
         expected: (#<Double "span_name">, {})
              got: (#<Double "span_name">)
       Diff:
       @@ -1 +1 @@
       -[#<Double "span_name">, {}]
       +[#<Double "span_name">]
```

There seems to be a weird incompatibility between RSpec and Ruby 2.7.

To solve this, I instead improved the test by passing an actual
argument. This works around the RSpec incompatibility and actually
improves coverage because previously we didn't actually test that
the `options` were correctly propagated (e.g. if we changed the
production code to set the options to always empty, the tests would
still pass).