Add profiler helper for detecting if we're on the main Ractor
**What does this PR do?**:

Add a new helper to `private_vm_access.c` that allows the caller
thread to ask if it's running as part of the main Ractor.

On Rubies without Ractor support, as a simplification, we always
return `true` so that callers don't need to probe for Ractor support.

**Motivation**:

This helper will be used to make sure the profiler code does not
run on the wrong Ractor, since the profiler code is currently not
built to be called in parallel from multiple Ractors.

**Additional Notes**:

Hey, I added the first spec that uses Ractors to our test suite!

**How to test the change?**:

Test coverage included.