Disable GC profiling on Ruby 3 by default due to Ruby VM Ractor-related bug
**What does this PR do?**:

This PR:

* Adds a toggle that can be used to enable/disable the GC profiling
  feature added in #2336
* Enables the GC profiling feature by default on Ruby 2.x and
  disables the feature by default on Ruby 3.x
* Exposes a new `DD_PROFILING_FORCE_ENABLE_GC` /
  `settings.profiling.advanced.force_enable_gc_profiling` setting to
  allow customers that don't use Ractors to manually re-enable the
  GC profiling feature.

**Motivation**:

While validating that the new profiler correctly works with Ractors,
I discovered that adding specs that invoked Ractors caused the
previously-built GC profiling feature (and attached specs) to start
failing.

I was able to track this down to an actual VM bug, which I've
reported upstream as <https://bugs.ruby-lang.org/issues/19112>.

To avoid impacting any customers (specifically, they would
potentially observe profiles with incomplete data), I decided to
disable GC profiling for Ruby 3.x customers for now.

I am hoping that the Ruby core team can be convinced to fix this
issue and backport it to existing Ruby 3.x stable releases, so
that we can again re-enable this feature for these users.

As an alternative, if this issue sticks around for a long time and
we really want this feature to be on by default for Ruby 3.x
customers, there are probably are ways for us to automatically
detect if Ractors are in use or not. I chose not to pursue such
a mechanism for now.

**Additional Notes**:

(None)

**How to test the change?**:

The change includes test coverage. You can also observe this
feature by profiling a Ruby application:

* On Ruby 2.x you will see garbage collection frames
* On Ruby 3.x you will not see garbage collection frames unless you
  force enable the feature.