Add after fork handling to new profiler codepath
In the previous commit, we refactored the old profiler codepath so that
the `Profiler` would call `#reset_after_fork` on collectors before
restarting them in a forking process.

In this commit, we use the added hook to propagate the
`#reset_after_fork` call to every component, so that they can clean up
their internal state in the child process of a fork.

In particular:
* The chain starts on the `Collectors::CpuAndWallTimeWorker` which must
  start by disabling its active `TracePoint`s so that something like
  GC doesn't trigger a sample while we're still resetting the profiler.

* Then the `Collectors::CpuAndWallTime` resets:
  a. Its per-thread tracking information. The native thread ids and
     CPU-time tracking of the thread that survived the fork need to
     be reset + all other threads that did not survive the fork need
     to be cleared.
  b. The internal stats need to be reset as well

* Then the `StackRecorder` resets:
  a. The active slot and its concurrency control. This is to avoid any
     issues if the fork happened while a serialization attempt was
     ongoing and left the concurrency control in an inconsistent
     state.
  b. The profiles, so there's no left over data from the parent
     process in the child process profiles.

The `#reset_after_fork` approach actually made the
`StackRecorder#clear` method added recently in #2362 actually not
be needed anymore, so I went ahead and removed it.