Make sure we reset the next profile while the GVL is being held
While working on validating that the new profiler is in good shape for
Ruby apps that use `fork`, I observed that it's not safe to call
`ddog_Profile_reset` without the Global VM Lock being held because
that means it can be "interrupted" by a VM `fork` and left in an
inconsistent state.

To avoid this, I've moved the reset operation to be performed before
we release the Global VM Lock.