[PROF-6463] Add `record_endpoint` to `StackRecorder`
**What does this PR do?**

This PR adds a new `record_endpoint` function to the `StackRecorder`,
which calls libdatadog's `set_endpoint` feature.

The new function is going to be called by the sampler thread and thus
follows the concurrency protocol for that thread, similar to the
existing `record_sample` function -- it locks the active profile before
working, and unlocks it after it is done.

Inside libdatadog, `set_endpoint` associates a given endpoint to
all samples that have a matching `local root span id`.

**Motivation**

This feature will be used to implement the endpoint profiling
feature on the new Ruby profiler. This will come as a later PR.

**Additional Notes**

The specs for this change also test the libdatadog behavior, since
the `StackRecorder` doesn't need to do much.

Nevertheless, I though it was important to have the extra coverage,
given the importance of this feature and that there are changes
planned on libdatadog that may affect it.

**How to test the change?**

The change includes test coverage. The later PR that adds support
for calling this API during sampling will also extend that test
coverage further.