Refactor environment variables in docker-compose using anchors
There's quite a lot of repetition here. I was about to add a new
environment variable, but before doing so, I decided to refactor this
a bit to avoid more copy-pasting.

We could actually use the same technique in more places in our
docker-compose.yml, but I decided to go with a smaller change for now.