[PROF-6566] Fix ddtrace installation issue when users have `CI=true`
**What does this PR do?**:

In #2358 we changed how the profiling native extension is built in CI
by adding the `-Werror` compiler option to turn warnings into errors.

We did this automatically by checking if the `CI` environment variable
was set to `true`.

Unfortunately, this broke at least one user, see
<https://github.com/DataDog/datadog-api-client-ruby/pull/1137>.
The warning that was triggered in that case was actually not
relevant, see my investigation in
<https://github.com/DataDog/dd-trace-rb/issues/2377>.

The `CI=true` was never meant to trigger outside of our CI, but
clearly it's too generic of a name for us to rely on that.

To fix this, I've gone ahead and changed the configuration so that
`-Werror` only gets added when `DD_PROFILING_CI=true`. I also
changed our docker and CircleCI configurations to set this by
default.

**Motivation**:

This issue affected one Datadog internal customer --
<https://github.com/DataDog/datadog-api-client-ruby/pull/1137>.

This should never happen -- we don't want any issues when
compiling the profiling bits to affect ddtrace installation.

**Additional Notes**:

(Nothing)

**How to test the change?**:

Validate that `-Werror` only gets added when
`DD_PROFILING_CI=true`.

To trigger a warning for testing, try removing
`DDTRACE_UNUSED` from function argument declarations, and you
should see that the compiler starts failing due to that.

---

Fixes #2377