Introduce binary version in CI cache key
Because CIRCLE_CACHE_VERSION is globally defined in CircleCI, bumping it
means starting from a blank universe for all subsequent job runs.

If a new PR that breaks binary compatibility ran + saved to cache first
then it would produce a matchable key from which other branches would
try to build, and fail for older jobs.

This also serves as a signal to developers that if they pull new images
past a change to this file they should clear their bundle cache volumes
and run bundle install and appraisal install again.