Add Ruby 2.5 support for `gvl_owner` function
There's a pretty big issue with `gvl_owner` on 2.5 and below (see
comments), which I'm not sure there's a reasonable solution for.

I still implemented the potentially-flawed `gvl_owner` as some
customers may still want to opt in to it (for instance in a
less-important environment, such as in staging/pre-prod).

I will also add a noisy warning about this in the logs, and plan to
document this in the known issues when we release beta (and GA) of the
new Ruby profiler.