Fix flaky spec
In <https://github.com/DataDog/dd-trace-rb/commit/e7e85a55491e21592a9ad7dc575c0644afcec01c>
as part of the PR that introduced GC profiling (#2336) we changed this
spec so that it expected non-GC samples.

But this test started causing flakyness on Ruby 2.3:
<https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/8009/workflows/65a15b34-90f1-466c-bbdc-4a5ccf9e9ebf/jobs/297416>

The source of the flakyness is that our `try_wait_until` only waited
for ANY sample to be available. Thus, if we get a profile with
ONLY GC samples, then the `try_wait_until` would finish and return
those.

Then, we'd filter the GC samples out and be left with nothing,
making the assertion fail.

To fix this, I've moved the GC samples filtering inside the
`try_wait_until` -- this way we know that it will either get
non-GC samples OR hit the timeout.

Finally, for some temporary insanity reason, when I updated this
spec for GC I called the local variable `current_thread_gc_samples`
where it should actually have been `current_thread_non_gc_samples`
or something like that.

This had me confused for a bit and I needed to reverse-engineer
my own oops ;)