Remove unneeded kill operation
In the context of the worker thread, calling `kill` is effectively
the same as calling `_native_stop`, so we don't need to do both.

I left only `_native_stop` because `kill` would skip over any code
after `_native_sampling_loop` inside `#start`, which is confusing
and makes it harder to test in some situations.