Fix flaky spec failing due to invalid deadline
The `pthread_cond_timedwait` API is a bit bizarre since instead of
getting a timeout it gets a wall-clock deadline (which is why we need
to `clock_gettime` first).

Previously, we were just blindly adding `10ms` to the `tv_nsec` field
but that was actually not correct BECAUSE this field can never be
>= 1 second. In this situation, the field is considered invalid and
`pthread_cond_timedwait` will return immediately with `EINVAL`.

This was, of course, causing flakyness in the test: if you ran the
spec too close to the second mark, e.g. 10:15:53.991, then
the loop could go through all iterations very quickly because
`pthread_cond_timewait` was returning immediately with an error,
causing the test to fail because the background thread did not get
a chance to run.

I've also bumped the number of iterations to 500 (times 10ms => 5s
deadline) just in case the test machine is a bit slower.