Directly store pointer to `CpuAndWallTimeWorker` state into global
Up until now, we were only storing the `active_sampler_instance`, e.g.
the Ruby `VALUE` from which we can extract the `CpuAndWallTimeWorker`
state.

But... this got really awkward, and potentially more expensive, as in
a bunch of places on the `CpuAndWallTimeWorker` we keep needing to
do `rb_typeddata_is_kind_of` and `TypedData_Get_struct`.

To simplify this, let's instead store the pointer to the state as well,
and use it directly when we need to.