First attempt at sampling apps that are sleeping (flawed)
My approach here was that when there's no thread holding the GVL,
then the sampling thread itself grabs the Global VM Lock and
triggers a sample.

This... mostly worked, but caused another issue: what if suddenly
the app is super busy and the sampling thread actually needs to wait
for the GVL? This may mean that we won't get any samples until the
Ruby scheduler graces the worker thread with the GVL, which on a busy
app can take many hundreds of milliseconds and breaks our approach
of actually interrupting Ruby regularly to sample.

I actually spotted this because one of the `CpuAndWallTimeWorker`
unit tests started failing in a flaky way: the
'is able to sample even when the main thread is sleeping' spec.

This was due to a race: as the profiler tried to sample, no thread
was holding the GVL. Then the background busy thread actually
grabbed the GVL and started its loop, and because it's busy, it will
hold the GVL for the usual scheduler time period (100ms), and thus
the profiler will not sample during that time.

The fix here will be to not block the worker thread while trying to
get the GVL, but introduce an additional thread that will "take
the fall" in case there's actually any blocking needed.

(Or, alternatively, if there was a "try-obtain-gvl-but-don't-block
for-it", this would solve the problem as well -- we just want some
Thread to be holding the GVL, doesn't matter which one. But
I could not find such an API and although I think it's not
impossible to build it, it would rely on a lot of internal knowledge
which doesn't seem like a great idea).

I've decided to commit the work so far as a checkpoint for later
reference anyway.