Introduce `StackRecorder#serialize!` to simplify testing
In a lot of tests we have a pattern of "I want a serialized output and
if something went wrong I just want to blow up".

We kept copy-pasting the same pattern to get this, and it got a bit
repetitive, so I decided to introduce a helper.

I could've added the helper somewhere else, but since it was so small
I decided to add it on the class itself.