Fix unsafely raising exceptions without global VM lock in new profiler
**What does this PR do?**:

While working on other parts of the profiler, I observed that if the
`serializer_flip_active_and_inactive_slots` function in
`stack_recorder.c` tried to raise an exception, the Ruby VM could
crash.

**Fortunately, the `StackRecorder` is only used on the new profiler
codepath, which is still considered in alpha state and off by default**.
Furthermore, the exceptions would only happen in cases that should not
happen in production.

The crash was caused by the fact that
`serializer_flip_active_and_inactive_slots` is called without the
Global VM Lock being held, and it's not valid to try to raise
exceptions in this state.

To fix this, and because I still find raising exceptions to be useful
in the native parts of the profiler, I've introduced two new helpers

* `grab_gvl_and_raise`
* `grab_gvl_and_raise_syserr`

that can be used to safely raise exceptions by making sure they
grab the GVL before calling the Ruby exception-raising APIs.

**Motivation**:

The added helpers may seem slightly overkill as they're currently
only used by the `StackRecorder`, but I plan to use them right
away in the implementation of a new component that's living in
a different branch.

**Additional Notes**:

(None)

**How to test the change?**:

Change includes test coverage.

Additionally, if you're curious to see the VM crashing, run the
`stack_recorder_spec.rb` without this fix and add

```ruby
Thread.new do
  Object.new while true
end
```

to the very top of the spec file. This will usually trigger the
crash because there will be two threads trying to allocate objects
in the Ruby heap -- the added one, as well as the thread running
`serializer_flip_active_and_inactive_slots`, which should set
off the fireworks.