Fix compiler warning by adding suggested attribute
On really modern versions of GCC (used in our Ruby 3.1 image), I
got the following cryptic warnings:

```
../../../../ext/ddtrace_profiling_native_extension/ruby_helpers.c: In function 'grab_gvl_and_raise':
../../../../ext/ddtrace_profiling_native_extension/ruby_helpers.c:48:3: error: function 'grab_gvl_and_raise' might be a candidate for 'gnu_printf' format attribute [-Werror=suggest-attribute=format]
   48 |   vsnprintf(args.exception_message, MAX_RAISE_MESSAGE_SIZE, format_string, format_string_arguments);
      |   ^~~~~~~~~
../../../../ext/ddtrace_profiling_native_extension/ruby_helpers.c: In function 'grab_gvl_and_raise_syserr':
../../../../ext/ddtrace_profiling_native_extension/ruby_helpers.c:72:3: error: function 'grab_gvl_and_raise_syserr' might be a candidate for 'gnu_printf' format attribute [-Werror=suggest-attribute=format]
   72 |   vsnprintf(args.exception_message, MAX_RAISE_MESSAGE_SIZE, format_string, format_string_arguments);
      |   ^~~~~~~~~
```

TL;DR as explained by
<https://www.ibm.com/docs/en/ztpf/2019?topic=warnings-function-might-be-candidate-format-attribute>
what GCC is telling us is that both `grab_gvl_and_raise` and
`grab_gvl_and_raise_syserr` are going to call a printf-style
function with their arguments, and that we can add a hint to
the compiler that tells it to validate such calls.

E.g. before this change doing
`grab_gvl_and_raise(rb_eRuntimeError, "Hello %d", "World")` would not
cause an error, and after this change the compiler will notice that
`%d` expects a number and not a string.