Add new `IdleSamplingHelper` component
Currently, the `CpuAndWallTimeWorker` has a big constraint: it can only
sample threads that have the global VM lock.

But what about when NO thread has the global VM lock? Currently, in
this situation, we still want to be able to take samples (e.g. to get
accurate wall-time), but we can't if no thread has the global VM lock.

In an earlier attempt, the `CpuAndWallTimeWorker` background thread
tried itself to grab the global VM lock, and then force a sample.
This mostly worked, but suffered from one big issue: what if suddenly
the VM became busy? In that situation, the `CpuAndWallTimeWorker`
suddenly would be "at the mercy" of the Ruby scheduler, and wait
a long time to get a sample vs setting its own pace and interrupting
Ruby as needed.

To solve this issue, this commit introduces a new component: the
`IdleSamplingHelper`. The `IdleSamplingHelper`'s only reason for
existance is to "take the hit" instead of the `CpuAndWallTimeWorker`.

When the `CpuAndWallTimeWorker` wants to take a sample and no thread
is holding the GVL, rather than it trying to directly acquire it,
it signals the `IdleSamplingHelper` to "get in line" for the Ruby
scheduler. This way, if suddenly the `IdleSamplingHelper` needs to
wait for a long time for the GVL, the `CpuAndWallTimeWorker` is
no longer affected, and can continue to sample the process.

This commit only adds the new component; a later commit will
wire it up.