Second attempt at sampling apps that are sleeping
This commit replaces the previous (flawed) approach of having the
`CpuAndWallTimeWorker` grab the Global VM Lock when it needed to take
a sample and the app was idle. This was problematic because Ruby may
concurrently get busy and so the `CpuAndWallTimeWorker` would be
blocked in line to acquire the GVL for an uncontrolled amount of time.

Instead, the `CpuAndWAllTimeWorker` asks the new `IdleSamplingHelper`
to do the same. The difference between both approaches is that
the `CpuAndWallTimeWorker` will always be free to trigger samples,
and the `IdleSamplingHelper` was explicitly designed to "take the fall"
and stand in queue for as long as needed.

Our tests did not need to be changed because the failure caused by
blocking was observed as a flaky test. With this fix, the
`cpu_and_wall_time_worker_spec.rb` should have no more flaky tests.