Remove flaky sanity checking in CpuAndWallTimeWorker
I've been running the `cpu_and_wall_time_worker_spec.rb` in a loop to
try to catch any remaining flakyness and identifed an issue with trying
to compare `trigger_sample_attempts` to the actual observed enqueued
samples (count did not match what the spec expected).

On a closer approach, I think the assertions being removed were
always misguided. In the common path, yeah, they would match the
expected counts but given that signal delivery is asynchronous, can be
lost, duplicate signals are ignored, and even that we now have the
`IdleSamplingHelper` in the mix, it means that we can't really rely
on `trigger_sample_attempts` being related to actually how many
times we sampled, because many things may cause attempts to be skipped.

I also think the other assertions left in those specs should provide
enough coverage, as they test what actually happened to the samples,
not sample /attempts/.